api/requirements.txt
beautifulsoup4==4.6.0
chardet==3.0.4
Flask==1.0.2
googlemaps==2.5.1
idna==2.6
IMDbPY==6.5
lxml==4.2.1
python-google-places==1.4.1
requests==2.18.4
selenium==3.11.0
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22